@@ -1,12 +1,17 @@
 import { TestBed } from '@angular/core/testing';
 
 import { OlympicService } from './olympic.service';
+import {provideHttpClient} from "@angular/common/http";
 
 describe('OlympicService', () => {
   let service: OlympicService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      providers: [
+        provideHttpClient()
+      ]
+    });
     service = TestBed.inject(OlympicService);
   });
 
