@@ -1,14 +1,19 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { HomeComponent } from './home.component';
+import {provideHttpClient} from "@angular/common/http";
 
-describe('HomeComponent', () => {
+describe('ChartTestComponent', () => {
   let component: HomeComponent;
   let fixture: ComponentFixture<HomeComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ HomeComponent ]
+      imports: [
+        HomeComponent
+      ],
+      providers: [
+        provideHttpClient(),
+      ]
     })
     .compileComponents();
 
